fix(fourslash,server-tests): split oversized shards (§19 cap)
The PR was growing two files already over the 2000-line §19 hard cap:
- handlers_code_fixes_tests.rs: 2879 → 2990 lines
- test-worker.cjs: 4623 → 4631 lines

Fix: move the new tests and parity additions out of the oversized shards.

handlers_code_fixes_tests.rs (2990 → 2879):
- Move the three nested-package.json tests into a new shard
  handlers_code_fixes_nested_pkg_tests.rs (153 lines), declared via
  #[path = "..."] mod nested_pkg_tests in handlers_code_fixes.rs.

test-worker.cjs (4631 → 4626):
- Extract the two parity-override conditions into a new
  import-fix-parity-overrides.cjs config file (20 lines).
- Replace the inline conditions + comments with a single
  importFixParityOverrides.some() call in both isImportFixParityTest
  and autoImportProviderParityTest.
